WT-3929 Fix "pip install" to work with Python3. - avoid the use of relative imports, it doesn't work well with   installed packages. - introduce packutil.py, to remove a circular dependency that causes   trouble with installed packages. - minor changes to setup_pip.py for Python3. - add a new library directory to check for standard compression libs.
